geotiff: extract attrs/metadata helpers to _attrs.py (#1883)
Step 5 of #1813's multi-PR refactor of __init__.py. Pure code motion;
no public API change.

Moved into a new xrspatial/geotiff/_attrs.py:

- _populate_attrs_from_geo_info: single source of truth for every
  read backend's DataArray.attrs (called 4x in __init__.py + every
  windowed-read path).
- _resolve_nodata_attr: nodata sentinel resolution from attrs
  (checks 'nodata', 'nodatavals', '_FillValue' in that order).
- _merge_friendly_extra_tags / _extract_rich_tags: write-side helpers
  that fold friendly attrs (image_description, extra_samples,
  colormap) into the extra_tags list passed to the writer.
- _extent_to_window: geographic extent to pixel-window math.

Plus the TIFF tag-id constants the helpers need (_TIFF_BYTE,
_TIFF_ASCII, _TIFF_SHORT, _RESOLUTION_UNIT_IDS) and the writer-side
constants used by to_geotiff (_LEVEL_RANGES, _VALID_COMPRESSIONS).

_extent_to_window is consumed by xrspatial/accessor.py via
``from .geotiff import _extent_to_window``; the re-import in
__init__.py keeps that path resolving.

Net __init__.py change: 4569 -> 4271 lines (-298). _attrs.py is 332
lines.

Verification: pixel-parity matrix from #1889 (192 cells), runtime
identity from #1890 (9 tests), attrs parity from #1548 (5 tests),
extra_tags filter (#1657), geotags, overview nodata inheritance
(#1739) -- 251/251 pass.

Refs #1813.
